Last week!
organized our branch a bit, so all the main info was right out front and we wouldn't have to go scrolling.
added my jupyter code, the updated R code, which is now a mess, don't use as a whole! and all the R photos that are in prezi. We got this!